move multiprocessing into backend
