+++ wandb/sdk_py27/backend/backend.py
@@ -20,13 +20,6 @@ from ..internal.internal import wandb_internal
 logger = logging.getLogger("wandb")
 
 
-if wandb.TYPE_CHECKING:  # type: ignore
-    from typing import (
-        Optional,
-        Union,
-    )
-
-
 class BackendThread(threading.Thread):
     """Class to running internal process as a thread."""
 
@@ -42,31 +35,49 @@ class BackendThread(threading.Thread):
 
 
 class Backend(object):
-    wandb_process = None
+    # multiprocessing context or module
+    # _multiprocessing: object
 
-    def __init__(self):
+    def __init__(self, settings=None, log_level=None):
         self._done = False
         self.record_q = None
         self.result_q = None
         self.wandb_process = None
         self.interface = None
         self._internal_pid = None
-        self._wl = wandb.setup()
+        self._settings = settings
+        self._log_level = log_level
+
+        self._multiprocessing = multiprocessing
+        self._multiprocessing_setup()
 
     def _hack_set_run(self, run):
         self.interface._hack_set_run(run)
 
-    def ensure_launched(
-        self,
-        settings=None,
-        log_level=None,
-        stdout_fd=None,
-        stderr_fd=None,
-        use_redirect=None,
-    ):
+    def _multiprocessing_setup(self):
+        if self._settings.start_method == "thread":
+            return
+
+        # defaulting to spawn for now, fork needs more testing
+        start_method = self._settings.start_method or "spawn"
+
+        # TODO: use fork context if unix and frozen?
+        # if py34+, else fall back
+        if not hasattr(multiprocessing, "get_context"):
+            return
+        all_methods = multiprocessing.get_all_start_methods()
+        logger.info(
+            "multiprocessing start_methods={}, using: {}".format(
+                ",".join(all_methods), start_method
+            )
+        )
+        ctx = multiprocessing.get_context(start_method)
+        self._multiprocessing = ctx
+
+    def ensure_launched(self):
         """Launch backend worker if not running."""
-        settings = dict(settings or ())
-        settings["_log_level"] = log_level or logging.DEBUG
+        settings = dict(self._settings or ())
+        settings["_log_level"] = self._log_level or logging.DEBUG
 
         # TODO: this is brittle and should likely be handled directly on the
         # settings object.  Multi-processing blows up when it can't pickle
@@ -74,10 +85,10 @@ class Backend(object):
         if "_early_logger" in settings:
             del settings["_early_logger"]
 
-        self.record_q = self._wl._multiprocessing.Queue()
-        self.result_q = self._wl._multiprocessing.Queue()
+        self.record_q = self._multiprocessing.Queue()
+        self.result_q = self._multiprocessing.Queue()
         if settings.get("start_method") != "thread":
-            process_class = self._wl._multiprocessing.Process
+            process_class = self._multiprocessing.Process
         else:
             process_class = BackendThread
             # disable interal process checks since we are one process
